Create crate for shared types
Some types must be available inside the `doco-derive` crate and its
macros, but cannot be exported by the crate itself. So we are adding a
new crate that implements the shared types and makes them available both
to `doco-derive` and users of doco.
