+++ crates/doco-types/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "doco-types"
+
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+getset = { workspace = true }
+typed-builder = { workspace = true }
