+++ crates/doco-types/src/test_utils.rs
@@ -1,0 +1,5 @@
+pub fn assert_send<T: Send>() {}
+
+pub fn assert_sync<T: Sync>() {}
+
+pub fn assert_unpin<T: Unpin>() {}
