+++ crates/doco/Cargo.toml
@@ -8,6 +8,7 @@ repository.workspace = true
 
 [dependencies]
 anyhow = "1.0.86"
+doco-types = { workspace = true }
 fantoccini = "0.21.2"
 getset = "0.1.2"
 reqwest = { version = "0.12.7", features = ["rustls-tls"] }
